+++ test/test_query.cpp
@@ -9917,7 +9917,7 @@ TEST(Query_ArrayLeafRelocate)
     }
 }
 
-TEST(Query_ColumnDeletion)
+TEST(Query_ColumnDeletionSimple)
 {
     Table foo;
     foo.add_column(type_Int, "a");
@@ -9938,6 +9938,85 @@ TEST(Query_ColumnDeletion)
     CHECK_LOGIC_ERROR(tv.sync_if_needed(), LogicError::column_does_not_exist);
     CHECK_EQUAL(x, 0);
     CHECK_EQUAL(tv.size(), 0);
+}
+
+TEST(Query_ColumnDeletionExpression)
+{
+    Table foo;
+    foo.add_column(type_Int, "a");
+    foo.add_column(type_Int, "b");
+    foo.add_column(type_Timestamp, "c");
+    foo.add_column(type_Timestamp, "d");
+    foo.add_empty_row(5);
+    foo.set_int(0, 0, 0);
+    foo.set_int(0, 1, 1);
+    foo.set_int(0, 2, 2);
+    foo.set_int(0, 3, 3);
+    foo.set_int(0, 4, 4);
+    foo.set_int(1, 0, 0);
+    foo.set_int(1, 1, 0);
+    foo.set_int(1, 2, 3);
+    foo.set_int(1, 3, 5);
+    foo.set_int(1, 4, 3);
+    foo.set_timestamp(2, 0, Timestamp(100, 100));
+    foo.set_timestamp(3, 0, Timestamp(200, 100));
+
+    auto q = foo.column<Int>(0) == foo.column<Int>(1) + 1;
+    TableView tv = q.find_all();
+    CHECK_EQUAL(tv.size(), 2);
+
+    foo.remove_column(0);
+    size_t x = 0;
+    CHECK_LOGIC_ERROR(x = q.count(), LogicError::column_does_not_exist);
+    CHECK_LOGIC_ERROR(tv.sync_if_needed(), LogicError::column_does_not_exist);
+    CHECK_EQUAL(x, 0);
+    CHECK_EQUAL(tv.size(), 0);
+
+    q = foo.column<Timestamp>(1) < foo.column<Timestamp>(2);
+    tv = q.find_all();
+    CHECK_EQUAL(tv.size(), 1);
+    foo.remove_column(2);
+    CHECK_LOGIC_ERROR(tv.sync_if_needed(), LogicError::column_does_not_exist);
+}
+
+TEST(Query_ColumnDeletionLinks)
+{
+    Group g;
+    TableRef foo = g.add_table("foo");
+    TableRef bar = g.add_table("bar");
+    TableRef foobar = g.add_table("foobar");
+
+    foobar->add_column(type_Int, "int");
+
+    bar->add_column(type_Int, "int");
+    bar->add_column_link(type_Link, "link", *foobar);
+
+    foo->add_column_link(type_Link, "link", *bar);
+
+    foobar->add_empty_row(5);
+    bar->add_empty_row(5);
+    foo->add_empty_row(10);
+    for (size_t i = 0; i < 5; i++) {
+        foobar->set_int(0, i, i);
+        bar->set_int(0, i, i);
+        bar->set_link(1, i, i);
+        foo->set_link(0, i, i);
+    }
+    auto q = foo->link(0).link(1).column<Int>(0) == 2;
+    auto q1 = foo->column<Link>(0).is_null();
+    auto tv = q.find_all();
+    auto cnt = q1.count();
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(cnt, 5);
+    // remove integer column, should not affect query
+    bar->remove_column(0);
+    tv.sync_if_needed();
+    CHECK_EQUAL(tv.size(), 1);
+    // remove link column, disaster
+    bar->remove_column(0);
+    CHECK_LOGIC_ERROR(tv.sync_if_needed(), LogicError::column_does_not_exist);
+    foo->remove_column(0);
+    CHECK_LOGIC_ERROR(q1.count(), LogicError::column_does_not_exist);
 }
 
 #endif // TEST_QUERY