+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed a bug in handover of detached linked lists. (issue #2378).
 * Fixed a few bugs (caused crashes) in pthread library on UWP.
+* Fixed a bug when deleting a column used in a query (#2408)
 
 ### Breaking changes